+++ gzap/apworld/ap_gzdoom_featured/__init__.py
@@ -1,0 +1,8 @@
+from worlds.AutoWorld import World
+
+class GZDoomExtrasWorld(World):
+    # Propitiate the world loader
+    game = "GZDoom Featured Maps"
+    hidden = True
+    item_name_to_id = {}
+    location_name_to_id = {}